Add csrf token endpoint
