+++ lib/interaction/commands/imihubun.js
@@ -1,36 +1,27 @@
-import * as D from "discord.js"
+export default async function imihubun(interaction, context) {
+  const { ALLOWED_CHANNEL_IDS, CHANNEL_COOLDOWN_MS, channelCooldowns, shiikurole, fetch } = context;
 
-const CHANNEL_COOLDOWN_MS = 60 * 1000; // 60秒
-const ALLOWED_CHANNEL_IDS = [
-  "123456789012345678", // #imihubun
-  "987654321098765432"
-];
-
-const channelCooldowns = new Map();
-
-export default async function imihubun(interaction) {
   await interaction.deferReply();
-    if (!interaction.member.roles.cache.has(prosess.env.shiikurole)) {
-      await interaction.editReply({
-        content: '❌ このコマンドを使用する権限がありません'
-      });
-      return;
-    }
+  if (!interaction.member.roles.cache.has(shiikurole)) {
+    await interaction.editReply({
+      content: '❌ このコマンドを使用する権限がありません'
+    });
+    return;
+  }
   let wordData = null;
   (async () => {
     const res = await fetch('https://povo-43.github.io/imihubun/words.json');
-     wordData = await res.json();
+    wordData = await res.json();
   })();
 
   const channel = interaction.options.getChannel('channel');
 
-    if (!channel.isTextBased()) {
-      await interaction.editReply({
-        content: '❌ テキストチャンネルを指定してください'
-      });
-      return;
-    }
-
+  if (!channel.isTextBased()) {
+    await interaction.editReply({
+      content: '❌ テキストチャンネルを指定してください'
+    });
+    return;
+  }
 
   /* ===== チャンネル制限 ===== */
   if (!ALLOWED_CHANNEL_IDS.includes(channel)) {
@@ -54,39 +45,33 @@ export default async function imihubun(interaction) {
 
   // クールダウン更新
   channelCooldowns.set(channel, now);
-  
-    // テキストチャンネル確認
 
-    const footer = "\n-# By [意味不文ジェネレーター](<https://povo-43.github.io/imihubun>)";
-    const main_text = Math.random() > 0.5
-      ? (
-          wordData.starts[Math.floor(Math.random() * wordData.starts.length)] +
-          wordData.subjects[Math.floor(Math.random() * wordData.subjects.length)] +
-          wordData.locations[Math.floor(Math.random() * wordData.locations.length)] +
-          wordData.actions[Math.floor(Math.random() * wordData.actions.length)] +
-          wordData.ends[Math.floor(Math.random() * wordData.ends.length)]
-        ) + ' ' + (
-          wordData.starts[Math.floor(Math.random() * wordData.starts.length)] +
-          wordData.subjects[Math.floor(Math.random() * wordData.subjects.length)] +
-          wordData.locations[Math.floor(Math.random() * wordData.locations.length)] +
-          wordData.actions[Math.floor(Math.random() * wordData.actions.length)] +
-          wordData.ends[Math.floor(Math.random() * wordData.ends.length)]
-        )
-      : (
-          wordData.starts[Math.floor(Math.random() * wordData.starts.length)] +
-          wordData.subjects[Math.floor(Math.random() * wordData.subjects.length)] +
-          wordData.locations[Math.floor(Math.random() * wordData.locations.length)] +
-          wordData.actions[Math.floor(Math.random() * wordData.actions.length)] +
-          wordData.ends[Math.floor(Math.random() * wordData.ends.length)]
-        )
-   
-    await channel.send(main_text + footer);
-  
-    await interaction.editReply({content: `✅ <#${channel.id}> に送信しました`});
-  }
+  // テキストチャンネル確認
 
-  if (!interaction.replied && !interaction.deferred) {
-  interaction.reply({ content: '❌ エラーが発生しました', flags: 64 })
-  .catch(console.error);
+  const footer = "\n-# By [意味不文ジェネレーター](<https://povo-43.github.io/imihubun>)";
+  const main_text = Math.random() > 0.5
+    ? (
+      wordData.starts[Math.floor(Math.random() * wordData.starts.length)] +
+      wordData.subjects[Math.floor(Math.random() * wordData.subjects.length)] +
+      wordData.locations[Math.floor(Math.random() * wordData.locations.length)] +
+      wordData.actions[Math.floor(Math.random() * wordData.actions.length)] +
+      wordData.ends[Math.floor(Math.random() * wordData.ends.length)]
+    ) + ' ' + (
+      wordData.starts[Math.floor(Math.random() * wordData.starts.length)] +
+      wordData.subjects[Math.floor(Math.random() * wordData.subjects.length)] +
+      wordData.locations[Math.floor(Math.random() * wordData.locations.length)] +
+      wordData.actions[Math.floor(Math.random() * wordData.actions.length)] +
+      wordData.ends[Math.floor(Math.random() * wordData.ends.length)]
+    )
+    : (
+      wordData.starts[Math.floor(Math.random() * wordData.starts.length)] +
+      wordData.subjects[Math.floor(Math.random() * wordData.subjects.length)] +
+      wordData.locations[Math.floor(Math.random() * wordData.locations.length)] +
+      wordData.actions[Math.floor(Math.random() * wordData.actions.length)] +
+      wordData.ends[Math.floor(Math.random() * wordData.ends.length)]
+    );
+
+  await channel.send(main_text + footer);
+
+  await interaction.editReply({content: `✅ <#${channel.id}> に送信しました`});
 }
-
